feat(front): The button to add to itinary is now a component
