+++ front/src/components/CardDetails.vue
@@ -87,9 +87,7 @@ export class Data implements CardData {
 </script>
 
 <script setup lang="ts">
-import { useMapPointsStore } from "@/stores/mapPoints";
-
-const mapPointsStore = useMapPointsStore();
+import PointButton from "@/components/PointButton.vue";
 
 const props = defineProps<{
   data: CardData;
@@ -128,24 +126,7 @@ const goback = (bool: boolean) => {
 
       <v-chip>Chip 3</v-chip>
     </v-chip-group>
-    <v-btn
-      block
-      size="large"
-      color="green"
-      prepend-icon="mdi-plus"
-      v-if="!mapPointsStore.includes($props.data)"
-      @click="mapPointsStore.add($props.data)"
-      >Ajouter à l'intinéraire</v-btn
-    >
-    <v-btn
-      block
-      size="large"
-      color="red"
-      prepend-icon="mdi-minus"
-      v-else
-      @click="mapPointsStore.remove($props.data)"
-      >Retirer de l'intinéraire</v-btn
-    >
+    <PointButton :data="props.data" :big="true"></PointButton>
     <v-divider></v-divider>
 
     <div v-if="props.data.description">
